+++ src/days/day_4/part2.rs
@@ -3,9 +3,9 @@ use std::io::{self, BufRead};
 use std::path::Path;
 
 
-pub fn run(file_name: &str) -> Result<u32, &'static str> {
+pub fn run(file_name: &str) -> Result<u64, &'static str> {
     let mut output = 0;
-    let mut cards = Vec::<(Vec<u32>, Vec<u32>)>::new();
+    let mut cards = Vec::<(Vec<u64>, Vec<u64>)>::new();
     let mut card_amounts = Vec::new();
 
     if let Ok(lines) = read_lines(file_name) {
@@ -51,7 +51,7 @@ pub fn run(file_name: &str) -> Result<u32, &'static str> {
     Ok(output)
 }
 
-fn parse_line(card_numbers: &mut Vec<u32>, winning_numbers: &mut Vec<u32>, line: &str) {
+fn parse_line(card_numbers: &mut Vec<u64>, winning_numbers: &mut Vec<u64>, line: &str) {
     let mut start = false;
     let mut parsed_cn = false;
 
@@ -66,7 +66,7 @@ fn parse_line(card_numbers: &mut Vec<u32>, winning_numbers: &mut Vec<u32>, line:
         }
 
         if c >= '0' && c <= '9' {
-            current_number = current_number * 10 + (c as u32 - '0' as u32);
+            current_number = current_number * 10 + (c as u64 - '0' as u64);
             continue;
         }
