+++ src/days/day_9/mod.rs
@@ -2,7 +2,7 @@ use crate::days::Part;
 use crate::days::{read_lines, parse_num};
 
 
-pub fn run(file_name: &str, part: Part) -> Result<u32, &'static str> {
+pub fn run(file_name: &str, part: Part) -> Result<u64, &'static str> {
     match part {
         Part::P1 => part1(file_name),
         Part::P2 => part2(file_name),
@@ -10,7 +10,7 @@ pub fn run(file_name: &str, part: Part) -> Result<u32, &'static str> {
 }
 
 
-fn part1(file_name: &str) -> Result<u32, &'static str> {
+fn part1(file_name: &str) -> Result<u64, &'static str> {
     let histories = parse_histories(file_name)?;
     let mut output = 0;
 
@@ -44,11 +44,11 @@ fn part1(file_name: &str) -> Result<u32, &'static str> {
 
     // println!("{}", output);
 
-    Ok(output as u32)
+    Ok(output as u64)
 }
 
 
-fn part2(file_name: &str) -> Result<u32, &'static str> {
+fn part2(file_name: &str) -> Result<u64, &'static str> {
     let histories = parse_histories(file_name)?;
     let mut output = 0;
 
@@ -83,7 +83,7 @@ fn part2(file_name: &str) -> Result<u32, &'static str> {
 
     println!("{}", output);
 
-    Ok(output as u32)
+    Ok(output as u64)
 }
